[refactor] enum 값 패키지 분리
