+++ core/data/src/main/java/com/ssavice/data/repositoryimpl/DemoSellerInfoRepository.kt
@@ -8,7 +8,7 @@ import com.ssavice.model.auth.CompanyVerifyToken
 import com.ssavice.model.seller.SellerMainInfo
 import com.ssavice.model.seller.SellerRegisterForm
 import com.ssavice.model.seller.SellerSummary
-import com.ssavice.model.service.ServiceState
+import com.ssavice.model.enums.ServiceState
 import com.ssavice.model.service.ServiceSummary
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
