+++ core/model/src/main/java/com/ssavice/model/enums/Category.kt
@@ -1,4 +1,4 @@
-package com.ssavice.model
+package com.ssavice.model.enums
 
 enum class Category(
     val value: String,
