+++ core/network/src/main/kotlin/com/ssavice/network/model/GetCompanyInfoDTO.kt
@@ -1,15 +1,12 @@
-import android.annotation.SuppressLint
-import android.app.Service
 import com.ssavice.model.Date
 import com.ssavice.model.Region
 import com.ssavice.model.seller.SellerMainInfo
-import com.ssavice.model.service.ServiceState
+import com.ssavice.model.enums.ServiceState
 import com.ssavice.model.service.ServiceSummary
-import com.ssavice.model.service.mapState
+import com.ssavice.model.enums.mapState
 import com.ssavice.network.model.RegionDTO
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import java.net.URL
 
 @Serializable
 data class GetCompanyInfoDTO(
