+++ feature/seller-main/src/main/kotlin/com/ssavice/seller_main/SellerMainUiState.kt
@@ -1,7 +1,6 @@
 package com.ssavice.seller_main
 
-import com.ssavice.model.service.ServiceState
-import java.net.URL
+import com.ssavice.model.enums.ServiceState
 
 sealed interface SellerMainUiState {
     data object Loading : SellerMainUiState
